Changes function name and moves to service package
Changed function name from `AddServiceToDevfile` to
`AddKubernetesComponentToDevfile`. Similar name change for delete
function. Also, moves it from `pkg/envinfo` to `pkg/service` package
